Add CLI scripts from osken to be installed via setup.cfg
Instead of keeping hardcoded files in bin/ directory it is
more flexible to install such files using setup.cfg file.
It will for example install files with correct shebang for
python2 and python3.

This patch also removes tools/osken-manager.spec file
as it was used by PyInstaller in Ryu and it isn't used
in os-ken.

Change-Id: Ie85c64d81c86f0059528ad4d8e238a4124d4e5b4