Fix pep8 errors in os-ken
Update the lower constraint for appdirs, similarly to how it has been
done in other repos.

Whitelist pep8 error E113.  This is the "unexpected indent" error.  This
module uses C-style comment blocks pervasively, things like:

    OFPP_IN_PORT = 0xfff8   # Send the packet out the input port. This
                            # virtual port must be explicitly used
                            # in order to send back out of the input
                            # port.
    OFPP_TABLE = 0xfff9     # Perform actions in flow table.
                            # NB: This can only be the destination
                            # port for packet-out messages.
    OFPP_NORMAL = 0xfffa    # Process with normal L2/L3 switching.

For the lines above the ones that have an indented comment with no code
to the left would get E113 with a recent change to a less permissive
policy.  It would cause over 260 changes to comment strings, destroying
the consistency of comment style, to make that change.  Therefore,
blacklist that pycodestyle error.

Change-Id: I6d2f463b50551f32bce863aa7dc0c70ad3054cbd
Story: #2005142