PDF documentation build
It turns out that blank lines are not allowed in tables in
the latex builder. Otherwise, LaTeX complication fails.
A blank line in the docstring of 'bfd' class in os_ken/lib/packet/bfd.py
was dropped to pass LaTeX build.

Note that the structure of the generated PDF is not good (bad enough?).
This will be updated later.

Change-Id: Iec62042d4642d50d411478fadfa09e6a6bbf0a42
Story: 2006099
Task: 35137