Replace deprecated brctl with ip commands
The bridge-utils package has been deprecated for some time now [1] and
'brctl' does not exist on some more recent distros like Fedora 28.
Replace references to brctl with the proper ip commands.

Calls to "brctl show" are not being replaced with calls to "bridge link"
because the output format is very different and in testing some bridges
were not listed.  So the simpler method of consulting /sys/class/net is
used.

[1] https://lwn.net/Articles/703776/ for example

Change-Id: I8a4891bb7ed44dc25ba3f79d2ace6927313c17b3
Related-Bug: #1801919