Fix lower-constraints job
The summary of changes in lower-constraints.txt are as below.

* Address version mismatch of eventlet in requirements.txt and
  lower-constraints.txt. eventlet 0.21.0 is blocked in
  global-requirements.txt, so the min version is set to 0.22.0.
* pytz version in l-c is bumped to satisfy Babel requirement.
* Update the min version of msgpack to 0.5.0 as 0.3.0 is not
  available in PyPI. msgpack-python is dropped as it is not used.
* Drop linter-specific entries from lower-constraints.txt
* hacking in l-c is updated to match the version in test-requirements.

I also ran requirements-check script locally and detected
a lot of inconsistencies in requirements related files in os-ken.
Errors detected by the script are fixed in this commit too.

NOTE: The above inconsistencies were not detected mainly because
we don't have requirements-check job in os-ken CI.
Perhaps it is due to 'tinyrpc' which is not included in g-r.
It is better to enable requirements-check job in os-ken CI,
but we need to add tinyrpc to global-requirements.txt or drop
tinyrpc dependencies (for example, by dropping os-ken/app
as tinyrpc is used only in os-ken/app).

Change-Id: I19617aee765f9086fbfdb67fb17839b83d9df48c