Drop lower-constraints.txt and its testing
As discussed in TC PTG[1] and TC resolution[2], we are
dropping the lower-constraints.txt file and its testing.
We will keep lower bounds in the requirements.txt file but
with a note that these are not tested lower bounds and we
try our best to keep them updated.

NOTE(elod.illes): setting neutron-tempest-plugin-dynamic-routing as
non-votin for now to unblock the gate, as the job is failing. This can
be set back as voting if a fix is found to this issue.

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal

Change-Id: I2a42a907170a6fcd8cae245f864121fd12422ef0
(cherry picked from commit 712535f6bec5cf40714c58fbb8924db3322b132c)
(cherry picked from commit 35df49c4aceae16a2758fd90193d56ee2a307607)