Fix quagga container
The quagga container failed to work because docker now seems to reset
the permissions of the /run folder on starting. The solution is to
set the permissions in the startup script.

Use the opportunity to switch to a newer base container. We cannot go
beyond ubuntu:20.04 though, since after that quagga has been replaced by
frr, which will require some deeper changes.

Change-Id: I6bbd88c0d66bffa0a310364e26cd73286524fee3