hub: prepare support for native python libraries
In this commit we moved StreamServer, StreamClient and LoggingWrapper
to the global scope and removed direct dependencies with eventlet when
needed.
The aim is to start separating logic unrelated to eventlet
implementation.

Also, imported module ssl, socket, traceback, and sys at the global
level to align with the new structure.

The change also brings unit tests for eventlet and native

Related-bug: #2087939
Change-Id: I7ed673fa9ceb4fb348c1c9e0c0bb001e6a331cf6
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@industrialdiscipline.com>