Remove CLI parts from os-ken
It has been agreed on one of the Neutron team meetings [1] that
we should remove those parts of os-ken which provides CLI scripts
and make it as small as possible to just fit Neutron actual needs.

The CLI bits were simply there from when os-ken was forked from Ryu
but we actually never used those tools in OpenStack so it is better
to remove them and keep os-ken as small as possible.

[1] https://meetings.opendev.org/meetings/networking/2025/networking.2025-02-25-14.00.log.html#l-81

Closes-bug: #2100268

Change-Id: I40a60a3233718fcda2f0be25120ef92708ffcbd5
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>